+++ packages/crashlytics/CHANGELOG.md
@@ -3,6 +3,48 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [8.1.2](https://github.com/invertase/react-native-firebase/compare/@react-native-firebase/crashlytics@8.1.1...@react-native-firebase/crashlytics@8.1.2) (2020-07-09)
+
+**Note:** Version bump only for package @react-native-firebase/crashlytics
+
+## [8.1.1](https://github.com/invertase/react-native-firebase/compare/@react-native-firebase/crashlytics@8.1.0...@react-native-firebase/crashlytics@8.1.1) (2020-07-07)
+
+**Note:** Version bump only for package @react-native-firebase/crashlytics
+
+# [8.1.0](https://github.com/invertase/react-native-firebase/compare/@react-native-firebase/crashlytics@8.0.3...@react-native-firebase/crashlytics@8.1.0) (2020-07-07)
+
+### Features
+
+- **android,ios:** upgrade native SDK versions ([#3881](https://github.com/invertase/react-native-firebase/issues/3881)) ([6cb68a8](https://github.com/invertase/react-native-firebase/commit/6cb68a8ea808392fac3a28bdb1a76049c7b52e86))
+
+## [8.0.3](https://github.com/invertase/react-native-firebase/compare/@react-native-firebase/crashlytics@8.0.2...@react-native-firebase/crashlytics@8.0.3) (2020-07-05)
+
+**Note:** Version bump only for package @react-native-firebase/crashlytics
+
+## [8.0.2](https://github.com/invertase/react-native-firebase/compare/@react-native-firebase/crashlytics@8.0.1...@react-native-firebase/crashlytics@8.0.2) (2020-07-05)
+
+**Note:** Version bump only for package @react-native-firebase/crashlytics
+
+## [8.0.1](https://github.com/invertase/react-native-firebase/compare/@react-native-firebase/crashlytics@8.0.0...@react-native-firebase/crashlytics@8.0.1) (2020-06-30)
+
+**Note:** Version bump only for package @react-native-firebase/crashlytics
+
+# [8.0.0](https://github.com/invertase/react-native-firebase/compare/@react-native-firebase/crashlytics@7.1.7...@react-native-firebase/crashlytics@8.0.0) (2020-06-30)
+
+- feat(crashlytics)!: upgrade to new Firebase Crashlytics SDK (#3580) ([cad58e1](https://github.com/invertase/react-native-firebase/commit/cad58e178b43dea461e17fa4a0a3fecd507ba68a)), closes [#3580](https://github.com/invertase/react-native-firebase/issues/3580)
+
+### BREAKING CHANGES
+
+- This is a breaking change to remove the use of the Fabric SDKs.
+
+## [7.1.7](https://github.com/invertase/react-native-firebase/compare/@react-native-firebase/crashlytics@7.1.6...@react-native-firebase/crashlytics@7.1.7) (2020-06-26)
+
+**Note:** Version bump only for package @react-native-firebase/crashlytics
+
+## [7.1.6](https://github.com/invertase/react-native-firebase/compare/@react-native-firebase/crashlytics@7.1.5...@react-native-firebase/crashlytics@7.1.6) (2020-06-22)
+
+**Note:** Version bump only for package @react-native-firebase/crashlytics
+
 ## [7.1.5](https://github.com/invertase/react-native-firebase/compare/@react-native-firebase/crashlytics@7.1.4...@react-native-firebase/crashlytics@7.1.5) (2020-06-10)
 
 **Note:** Version bump only for package @react-native-firebase/crashlytics